requirements.txt
python-dotenv>=1.0.1
discord.py>=2.4.0
ffmpeg>=1.4
yt-dlp>=2023.12.13
ping3>=4.0.8
discord-ext-voice-recv>=0.4.2a145
colorlog>=6.9.0
PyYAML>=6.0.2
fastapi>=0.115.7
Jinja2>=3.1.5
python-multipart>=0.0.6
uvicorn>=0.34.0
scikit-learn>=1.6.1
gTTS>=2.5.4
aiohttp>=3.9.0
requests>=2.31.0
psutil>=7.0.0
pydantic>=2.10.6
SQLAlchemy>=2.0.38
mysql>=0.0.3
mysql-connector-python>=9.2.0
pytest>=8.3.4
PyMySQL>=1.1.1
pathlib>=1.0.1